fixes + spostato comandi per send/recv
